requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==23.2.0
av==9.2.0
black==25.1.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.15.1
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.8
cloudpickle==3.1.1
cmake==3.29.2
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@9604f5995cc628619f0e4fd913453b4d7d61db3f
distlib==0.3.8
exceptiongroup==1.2.1
expecttest==0.2.1
filelock==3.14.0
fonttools==4.56.0
frozenlist==1.4.1
fsspec==2024.3.1
fvcore==0.1.5.post20221221
google-auth==2.29.0
google-auth-oauthlib==0.4.6
grad-cam==1.5.4
grpcio==1.63.0
habana-media-loader==1.14.0.493
habana-pyhlml==1.14.0.493
habana-torch-dataloader @ file:///tmp/tmp.yp35aLg1m9/habana_torch_dataloader-1.14.0.493-cp310-cp310-linux_x86_64.whl#sha256=d57c0e52bf97b9a38a261986ed34d1ed59986fccd0d48d8ca15712221855640e
habana-torch-plugin @ file:///tmp/tmp.yp35aLg1m9/habana_torch_plugin-1.14.0.493-cp310-cp310-linux_x86_64.whl#sha256=a342bf1183f7813d2ddfb893370bf10a06bd4490ac435d6c1e262a096f1986a4
habana_gpu_migration @ file:///tmp/tmp.yp35aLg1m9/habana_gpu_migration-1.14.0.493-cp310-cp310-linux_x86_64.whl#sha256=dd08b8a0b53571b9f9019cddf1ba38d53312f200c2cd7de66b7185ea5c6cccc2
habana_quantization_toolkit @ file:///tmp/tmp.yp35aLg1m9/habana_quantization_toolkit-1.14.0.493-py3-none-any.whl#sha256=32bf985b89ca80889442ce2961f2ec831f1352fdbff34bc0089bcb48f47f8809
huggingface-hub==0.29.3
hydra-core==1.3.2
identify==2.5.36
idna==3.7
iniconfig==2.0.0
intel-openmp==2023.2.4
iopath==0.1.9
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.8
lightning==2.1.2
lightning-habana==1.3.0
lightning-utilities==0.11.2
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.10.1
mkl==2023.1.0
mkl-include==2023.1.0
mpi4py==3.1.4
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
ninja==1.11.1.1
nodeenv==1.8.0
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==24.0
pandas==2.0.1
pathspec==0.12.1
perfetto==0.7.0
pillow==11.1.0
Pillow-SIMD==7.0.0.post3
platformdirs==4.2.1
pluggy==1.5.0
portalocker==3.1.1
pre-commit==3.3.3
pretty-errors==1.2.25
protobuf==3.20.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybind11==2.10.4
pycocotools==2.0.8
pycparser==2.22
pydantic==1.10.13
pynvml==8.0.4
pyparsing==3.2.1
pytest==8.2.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.4
pytz==2024.1
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.16.0
sympy==1.12
tabulate==0.9.0
tbb==2021.12.0
tdqm==0.0.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.5.0
threadpoolctl==3.6.0
timm==1.0.15
tomli==2.0.1
torch @ file:///tmp/tmp.yp35aLg1m9/torch-2.1.1a0%2Bgitb51c9f6-cp310-cp310-linux_x86_64.whl#sha256=1abf98885ccc265886480bdc3e26f3b7eebf19d0e7913eb75e2ad980b7d70089
torch_tb_profiler @ file:///tmp/tmp.yp35aLg1m9/torch_tb_profiler-0.4.0-py3-none-any.whl#sha256=0d3af22de662e6641215b5e7cd2b3472d4ef2c4fa90a6b5ae43fcca72301db7d
torchaudio @ file:///tmp/tmp.yp35aLg1m9/torchaudio-2.1.0%2B6ea1133-cp310-cp310-linux_x86_64.whl#sha256=d32495f49785a114acdeb2299c9006015b9d7b0f2c4c5ba81908dc35ae09d237
torchdata @ file:///tmp/tmp.yp35aLg1m9/torchdata-0.7.0%2Bc5f2204-py3-none-any.whl#sha256=a675577c0018ca609e5e21e0c6bc712e6aa3d1e119d9ffd2ec1a09194f8dae4e
torchmetrics==1.4.0
torchtext @ file:///tmp/tmp.yp35aLg1m9/torchtext-0.16.0a0%2B4e255c9-cp310-cp310-linux_x86_64.whl#sha256=4a373211b2f80e632aed4143f2789d7102516a61c69eaab9890814543022b192
torchvision @ file:///tmp/tmp.yp35aLg1m9/torchvision-0.16.0%2Bfbb4cc5-cp310-cp310-linux_x86_64.whl#sha256=273904fb11dacebc32e66e3a03a9d206fe61d3d45bad914f8e0eaf439f8f43fc
tqdm==4.66.4
ttach==0.0.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
virtualenv==20.26.1
Werkzeug==3.0.3
yacs==0.1.8
yamllint==1.35.1
yarl==1.9.4
